+++ source/ChanSort.Loader.SatcoDX/SatcoDxPlugin.cs
@@ -1,4 +1,6 @@
-﻿using ChanSort.Api;
+﻿using System.IO;
+using System.Text;
+using ChanSort.Api;
 
 namespace ChanSort.Loader.SatcoDX
 {
@@ -10,6 +12,14 @@ namespace ChanSort.Loader.SatcoDX
 
     public SerializerBase CreateSerializer(string inputFile)
     {
+      var buffer = new byte[7];
+      using (var strm = new FileStream(inputFile, FileMode.Open))
+      {
+        var len = strm.Read(buffer, 0, buffer.Length);
+        if (len != buffer.Length || Encoding.ASCII.GetString(buffer, 0, len) != "SATCODX")
+          return null;
+      }
+
       return new Serializer(inputFile);
     }
   }